extract util function
